refactoring bien avancé, début du readme
