--- App.js
@@ -1,26 +1,0 @@
-import React from 'react';
-import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator } from '@react-navigation/stack';
-import HomeScreen from './HomePage';
-import GamePage from './GamePage'; // Assurez-vous que GamePage est importé depuis le bon emplacement
-
-const Stack = createStackNavigator();
-
-export default function App() {
-    return (
-        <NavigationContainer>
-            <Stack.Navigator>
-                <Stack.Screen
-                    name="HomePage"
-                    component={HomeScreen}
-                    options={{ headerShown: false }}
-                />
-                <Stack.Screen
-                    name="GamePage"
-                    component={GamePage}
-                    options={{ headerShown: false }}
-                />
-            </Stack.Navigator>
-        </NavigationContainer>
-    );
-}
