+++ app.json
@@ -9,13 +9,18 @@
     "userInterfaceStyle": "automatic",
     "newArchEnabled": true,
     "ios": {
-      "supportsTablet": true
+      "supportsTablet": true,
+      "bundleIdentifier": "com.frixyo.wakwa",
+      "infoPlist": {
+        "ITSAppUsesNonExemptEncryption": false
+      }
     },
     "android": {
       "adaptiveIcon": {
         "foregroundImage": "./assets/images/adaptive-icon.png",
         "backgroundColor": "#ffffff"
-      }
+      },
+      "package": "com.frixyo.wakwa"
     },
     "web": {
       "bundler": "metro",
@@ -39,7 +44,7 @@
           "enableFTS": true,
           "useSQLCipher": true,
           "android": {
-            "enableFTS": true, 
+            "enableFTS": true,
             "useSQLCipher": true
           },
           "ios": {
@@ -53,6 +58,14 @@
     ],
     "experiments": {
       "typedRoutes": true
+    },
+    "extra": {
+      "router": {
+        "origin": false
+      },
+      "eas": {
+        "projectId": "a2b79af8-af2b-4200-98e4-228266e4d239"
+      }
     }
   }
 }
