+++ app/Model/AddPlayersProps.tsx
@@ -1,0 +1,5 @@
+import Joueur from "./Joueur";
+
+export default interface AddPlayerProps {
+    setJoueurs: (joueurs: Joueur[]) => void;
+}
