+++ app/Model/Case.tsx
@@ -1,0 +1,6 @@
+import { imageMapping } from '../Constant/images';
+
+export default interface Case {
+    description: string;
+    image: keyof typeof imageMapping;
+}
