+++ app/Model/EditProps.tsx
@@ -1,0 +1,3 @@
+export default interface EditProps {
+    plateauId: string;
+}
