+++ app/Model/GameProps.tsx
@@ -1,0 +1,3 @@
+export default interface GameProps {
+    plateauId: string;
+}
