+++ app/Model/Joueur.tsx
@@ -1,0 +1,6 @@
+import { playerImageMapping } from '../Constant/images';
+
+export default interface Joueur {
+    name: string;
+    image: keyof typeof playerImageMapping;
+}
