+++ app/Model/ModalActionGameProps.tsx
@@ -1,0 +1,7 @@
+import Case from "../Model/Case";
+
+export default interface ModalActionGameProps {
+    isModalVisible: boolean;
+    closePopupActionGame: () => void;
+    currentCase: Case;
+}
