+++ app/Model/ModalMainProps.tsx
@@ -1,0 +1,6 @@
+export default interface ModalMainProps {
+  isModalVisible: boolean;
+  closePopup: () => void;
+  selectedPlateauIndex: string | null;
+  listPlateaux: Record<string, number>;
+}
