+++ app/Model/ModalParametreProps.tsx
@@ -1,0 +1,4 @@
+export default interface ModalParametreProps {
+    isModalVisible: boolean;
+    closePopupParametre: () => void;
+}
