+++ app/Model/Plateau.tsx
@@ -1,0 +1,4 @@
+export default interface Plateau {
+    id: string;
+    name: string;
+}
