+++ app/Model/PlayersProps.tsx
@@ -1,0 +1,6 @@
+import Joueur from "./Joueur";
+
+export default interface PlayersProps {
+    joueurs: Joueur[];
+    setJoueurs: (joueurs: Joueur[]) => void;
+}
