+++ app/Page/EditPage.tsx
@@ -1,16 +1,19 @@
 import { View, StyleSheet, ImageBackground } from 'react-native';
 import { useLocalSearchParams } from "expo-router";
 import { SQLiteProvider } from 'expo-sqlite';
-import Edit from '../Component/Edit';
+
+// Import Component
+import Edit from '../Component/Core/Edit';
 
 export default function EditPage() {
   const { plateauId } = useLocalSearchParams();
+
   return (
     <SQLiteProvider databaseName="mydatabasetest3.db" assetSource={{ assetId: require('../../assets/db/mydatabasetest3.db') }}>
       <ImageBackground source={require('../../assets/bg/bg_1.png')} style={styles.background}>
-        <View>
-          <Edit plateauId={plateauId}/>
-        </View>
+
+        <Edit plateauId={Array.isArray(plateauId) ? plateauId[0] : plateauId}/>
+
       </ImageBackground>
     </SQLiteProvider>
   );
