+++ app/_layout.js
@@ -4,9 +4,9 @@ export default function RootLayout() {
   return (
     <Stack>
       <Stack.Screen name="index" options={{headerShown: false}} />
-      <Stack.Screen name="JoueurPage" options={{headerShown: false}} />
-      <Stack.Screen name="EditPage" options={{headerShown: false}} />
-      <Stack.Screen name="GamePage" options={{headerShown: false, gestureEnabled: true }} />
+      <Stack.Screen name="Page/JoueurPage" options={{headerShown: false}} />
+      <Stack.Screen name="Page/EditPage" options={{headerShown: false}} />
+      <Stack.Screen name="Page/GamePage" options={{headerShown: false, gestureEnabled: true }} />
     </Stack>
   );
 }
