--- app/images.js
@@ -1,24 +1,0 @@
-export const imageMapping = {
-    'carteDebut': require('../assets/cases/carteDebut.png'),
-    'cartebb_01': require('../assets/cases/cartebb_01.png'),
-    'cartemarron_01': require('../assets/cases/cartemarron_01.png'),
-    'carterouge_01': require('../assets/cases/carterouge_01.png'),
-    'cartevert_01': require('../assets/cases/cartevert_01.png'),
-    'carteviolet_01': require('../assets/cases/carteviolet_01.png'),
-    'carteFin': require('../assets/cases/carteFin.png'),
-    'cartevide_01': require('../assets/cases/cartevide_01.png')
-};
-
-export const imageMappingPlayable = {
-    'cartebb_01': require('../assets/cases/cartebb_01.png'),
-    'cartemarron_01': require('../assets/cases/cartemarron_01.png'),
-    'carterouge_01': require('../assets/cases/carterouge_01.png'),
-    'cartevert_01': require('../assets/cases/cartevert_01.png'),
-    'carteviolet_01': require('../assets/cases/carteviolet_01.png'),
-};
-
-export const playerImageMapping = {
-    'pionBleu': require('../assets/pion/pionBleu.png'),
-    'pionVert': require('../assets/pion/pionVert.png'),
-    'pionRose': require('../assets/pion/pionRose.png'),
-};
