+++ commande.md
@@ -1,0 +1,25 @@
+
+
+1. Install dependencies
+
+   ```bash
+      npm install
+   ```
+
+2. Start the app
+
+   ```bash
+      npm run start
+   ```
+
+## Reinstaller la BD
+
+   go into assets/db
+   delete mydatabase.db
+   ```bash
+      sqlite3 mydatabase.db < init.sql
+   ```
+3. Expo go
+
+   CTRL+C pour fermer l'appli
+
