+++ package.json
@@ -42,7 +42,8 @@
     "react-native-safe-area-context": "^4.12.0",
     "react-native-screens": "~4.4.0",
     "react-native-web": "~0.19.13",
-    "react-native-webview": "13.12.5"
+    "react-native-webview": "13.12.5",
+    "expo-dev-client": "~5.0.11"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
